% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_api.R
\name{call_reb_api}
\alias{call_reb_api}
\title{Title}
\usage{
call_reb_api(
  base_url,
  key = "sample key",
  Type = "xml",
  pIndex = 1,
  pSize = 5,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{
res1 <- get_api_key() |>
        reb_tbl_list(Type = "json", pIndex = 1, pSize = 5)

res1 |> parse_reb_api_json_context()

res1 |> parse_reb_api_json()


}
