% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_api_key}
\alias{get_api_key}
\title{Get API Key}
\usage{
get_api_key(key = NULL)
}
\arguments{
\item{key}{A character string representing the API key. If provided, this will set the REBONEKEY environment variable.}
}
\value{
A character string containing the API key.
}
\description{
Get API Key
}
